Rename Buildings_funcs_grid_rev2.py to buildings_funcs_grid.py; Refactor
