+++ message_ix_buildings/chilled/buildings_funcs_grid.py
